+++ packages/std/__tests__/types/isAsyncFunction.spec.ts
@@ -7,10 +7,10 @@ describe('Types > isAsyncFunction', () => {
     expect(isAsyncFunction(() => {})).toBe(false);
     expect(isAsyncFunction(() => {})).toBe(false);
     expect(isAsyncFunction(function* () {})).toBe(false);
+    expect(isAsyncFunction(async function* () {})).toBe(false);
   });
 
   test('valid cases', () => {
     expect(isAsyncFunction(async () => {})).toBe(true);
-    expect(isAsyncFunction(async function* () {})).toBe(true);
   });
 });