+++ packages/std/src/types/isAsyncFunction.ts
@@ -11,7 +11,6 @@
 
 export default function isAsyncFunction(
   val: unknown,
-): val is ((...args: unknown[]) => Promise<unknown>) | AsyncGeneratorFunction {
-  const a = ['[object AsyncFunction]', '[object AsyncGeneratorFunction]'];
-  return a.includes(Object.prototype.toString.call(val));
+): val is (...args: unknown[]) => Promise<unknown> {
+  return Object.prototype.toString.call(val) === '[object AsyncFunction]';
 }