+++ packages/std/src/types/isGeneratorFunction.ts
@@ -11,6 +11,5 @@
 export default function isGeneratorFunction(
   val: unknown,
 ): val is GeneratorFunction {
-  const a = ['[object GeneratorFunction]', '[object AsyncGeneratorFunction]'];
-  return a.includes(Object.prototype.toString.call(val));
+  return Object.prototype.toString.call(val) === '[object GeneratorFunction]';
 }